# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Future Jaano - AI-powered spiritual guidance platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /horoscope/daily:
    get:
      operationId: getDailyHoroscope
      tags: [horoscope]
      summary: Get daily horoscope for all zodiac signs
      parameters:
        - name: date
          in: query
          schema:
            type: string
          description: Date in YYYY-MM-DD format (defaults to today)
        - name: lang
          in: query
          schema:
            type: string
            enum: [en, hi]
          description: Language (en or hi)
      responses:
        "200":
          description: Daily horoscopes for all signs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HoroscopeEntry"
  /horoscope/daily/{sign}:
    get:
      operationId: getDailyHoroscopeBySign
      tags: [horoscope]
      summary: Get daily horoscope for a specific zodiac sign
      parameters:
        - name: sign
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Daily horoscope for the sign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HoroscopeEntry"
  /horoscope/weekly/{sign}:
    get:
      operationId: getWeeklyHoroscope
      tags: [horoscope]
      summary: Get weekly horoscope for a zodiac sign
      parameters:
        - name: sign
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Weekly horoscope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HoroscopeEntry"

  # KUNDLI
...
components:
  schemas:
...
    HoroscopeEntry:
      type: object
      required: [sign, date, prediction, luckyColor, luckyNumber, compatibility]
      properties:
        sign: { type: string }
        signHindi: { type: ["string", "null"] }
        date: { type: string }
        prediction: { type: string }
        health: { type: string }
        career: { type: string }
        love: { type: string }
        finance: { type: string }
        luckyColor: { type: string }
        luckyNumber: { type: integer }
        luckyGem: { type: ["string", "null"] }
        compatibility: { type: string }
        rating: { type: integer }
        emoji: { type: string }

    # KUNDLI
...